+++ .gitignore
@@ -1,5 +1,5 @@
 /node_modules
 .env
-
+src/public/img/uploads
 firebase.json
