+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "env-cmd --file ./.env nodemon index.js"
+    "dev": "env-cmd --file ./.env nodemon src/index.js"
   },
   "author": "Diego Montoya",
   "license": "ISC",
@@ -21,7 +21,8 @@
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
     "sharp": "^0.32.1",
-    "timeago.js": "^4.0.2"
+    "timeago.js": "^4.0.2",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.20"
