+++ src/controllers/MusicaController.js
@@ -1,28 +1,6 @@
 const { db } = require("../database/config");
 const { Timestamp } = require("firebase-admin/firestore");
-const multer = require('multer');
-const path = require('path');
-const sharp = require('sharp');
 
-//multer
-const storage = multer.diskStorage({
-  destination: (req, file, cb)=>{
-    cb(null, "./public/uploads");
-  },
-  filename: (req, file, cb)=>{
-    const ext = file.originalname.split(".").slice(-1)[0];
-    const parsedFileName = path.parse(file.originalname).name;
-    cb(null, `${file.fieldname}-${parsedFileName}-${Date.now()}.${ext}`);
-  }
-})
-
-//sharp
-
-const helperImg = (filePath, filename, size = 200)=>{
-  return sharp(filePath).resize(size).toFile(`./public/optimize/${filename}`)
-}
-
-const upload = multer({storage});
 
 const getAll = async (req, res) => {
     try {
@@ -40,19 +18,16 @@ const getAll = async (req, res) => {
 
 const create = async (req, res) => {
     try {
-      let uriFile = `http://${req.hostname}:${req.socket.localPort}/images/${req.file.filename}`
-      let resizePic = `http://${req.hostname}:${req.socket.localPort}/resource/resize-${req.file.filename}`
-      helperImg(req.file.path, `resize-${req.file.filename}`, 200)
       const agregar = db.collection("Musica").doc();
       await agregar.set({
         artistaBanda: req.body.artistaBanda,
         cancion: req.body.cancion,
         enlace: req.body.enlace,
-        imagen: resizePic,
+        imagen: req.file.path,
         fechaPost: Timestamp.now().toDate().toString(),
       });
-      res.status(200).send({ url: uriFile, message: "Element created succesfully"});
-      console.log(req.body, resizePic)
+      res.status(200).json(agregar.data);
+      
     } catch (error) {
       res.status(500).json({ error: `An error ocurred ${error}` }); 
       console.log(req.body )
@@ -91,4 +66,4 @@ const getById = async (req, res) => {
     }
   }
 
-  module.exports = {getAll, create, remove, edit, getById, upload}
+  module.exports = {getAll, create, remove, edit, getById}
