+++ src/database/config.js
@@ -9,6 +9,4 @@ initializeApp({
 
 const db = getFirestore();
 
-module.exports = {
-    db
-};
+module.exports = {db}
