+++ src/views/index.ejs
@@ -3,14 +3,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title><%= titulo %></title>
 </head>
 <body>
         <form action="/Musica" method="post" enctype="multipart/form-data">
             <input type="text" id="artistaBanda" name="artistaBanda">
             <input type="text" id="cancion" name="cancion">
             <input type="text" id="enlace" name="enlace">
-            <input type="file" id="pic" name="pic">
+            <input type="file" id="image" name="image">
             <button type="submit">Agregar</button>
         </form>
 </body>
